+++ src/utils/throttleAsync.ts
@@ -1,0 +1,45 @@
+import {wait} from '@feltcoop/felt';
+
+// TODO maybe support return values? gets tricky: what should it return for the skipped ones?
+
+/**
+ * Throttles calls to a promise-returning function.
+ * If the throttled function is called while the promise is pending,
+ * it's queued up to run after the promise completes,
+ * and only the last call is executed;
+ * calls except the most recent made during the pending promise are discarded.
+ * This is distinct from a queue where every call to the throttled function eventually runs.
+ * @param fn Any promise-returning function.
+ * @param toCacheKey Throttled calls are grouped by the `Map` key returned from this function.
+ * @param delay Throttled calls delay this many milliseconds after the previous call finishes.
+ * @returns Same as `fn`.
+ */
+export const throttleAsync = <TArgs extends any[]>(
+	fn: (...args: TArgs) => Promise<void>,
+	toCacheKey: (...args: TArgs) => any,
+	delay: number = 0,
+): ((...args: TArgs) => Promise<void>) => {
+	const cache: Map<string, {id: number; promise: Promise<void>}> = new Map();
+	let _id = 0;
+	return async (...args) => {
+		const id = _id++;
+		const cacheKey = toCacheKey(...args);
+		let cached = cache.get(cacheKey);
+		if (cached) {
+			cached.id = id; // queue this one up
+			await cached.promise;
+			if (cached.id !== id) return; // a later call supercedes this one
+			if (delay) await wait(delay);
+		}
+		const promise = fn(...args).then(() => {
+			if (id === cached!.id) {
+				cache.delete(cacheKey); // delete only when we're done with this `cacheKey`
+			}
+		});
+		if (!cached) {
+			cached = {promise, id};
+			cache.set(cacheKey, cached);
+		}
+		await promise;
+	};
+};